+++ cmake/CompilerWarnings.cmake
@@ -6,7 +6,6 @@
 
 function(set_project_warnings project_name)
   set(MSVC_WARNINGS
-      /W4     # Baseline reasonable warnings
       /w14242 # 'identifier': conversion from 'type1' to 'type1', possible loss
               # of data
       /w14254 # 'operator': conversion from 'type1:field_bits' to